feature: add python and agentops sdk version to hostenv
